tests/pdf_extraction/requirements.txt
pymupdf 
pytesseract 
pdf2image 
pillow